+++ eras/babbage/impl/src/Cardano/Ledger/Babbage/UTxO.hs
@@ -14,12 +14,15 @@ import Cardano.Ledger.Babbage.Era (BabbageEra)
 import Cardano.Ledger.Babbage.TxBody ()
 import Cardano.Ledger.Crypto (Crypto)
 import Cardano.Ledger.Mary.UTxO (getConsumedMaryValue)
+import Cardano.Ledger.Shelley.UTxO (shelleyProducedValue)
 import Cardano.Ledger.UTxO (EraUTxO (..))
 
 instance Crypto c => EraUTxO (BabbageEra c) where
   type ScriptsNeeded (BabbageEra c) = AlonzoScriptsNeeded (BabbageEra c)
 
   getConsumedValue = getConsumedMaryValue
+
+  getProducedValue = shelleyProducedValue
 
   getScriptsNeeded = getAlonzoScriptsNeeded