+++ libs/cardano-ledger-core/src/Cardano/Ledger/UTxO.hs
@@ -7,7 +7,6 @@
 {-# LANGUAGE FlexibleInstances #-}
 {-# LANGUAGE GeneralizedNewtypeDeriving #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
-{-# LANGUAGE PatternSynonyms #-}
 {-# LANGUAGE ScopedTypeVariables #-}
 {-# LANGUAGE StandaloneDeriving #-}
 {-# LANGUAGE TypeApplications #-}
@@ -54,6 +53,7 @@ import Cardano.Ledger.Crypto (Crypto)
 import Cardano.Ledger.Keys (
   DSignable,
   Hash,
+  KeyHash,
   KeyRole (..),
   verifySignedDSIGN,
  )
@@ -211,6 +211,13 @@ class EraTxBody era => EraUTxO era where
     -- | Function that can lookup current delegation deposits
     (StakeCredential (EraCrypto era) -> Maybe Coin) ->
     UTxO era ->
+    TxBody era ->
+    Value era
+
+  getProducedValue ::
+    PParams era ->
+    -- | Check whether a pool with a supplied PoolStakeId is already registered.
+    (KeyHash 'StakePool (EraCrypto era) -> Bool) ->
     TxBody era ->
     Value era